DERIVED_MV tool
In the derived_mv experiment, a new inter prediction mode is added where
the motion vector is derived by matching an "L" shaped region of
reconstructed pixels adjacent to the current block against the reference
frame. The MV is not transmitted in bitstream as the decoder can repeat
the process to derive the MV.

DERIVED_MV is enabled for: NEARMV, NEAR_NEARMV and skip modes, and is
designed to avoid parsing dependency.

When derived MV is used, We disable signaling motion_mode as simple
translation implicitly.