ESR: add syntax for 2d superres scaling factor
The new syntax supports 4 possible scaling factors, which will be
applied to both horizontal and vertical directions. The index is
first parsed, and the decoder will retrieve the scaling factor from
a lookup table. Currently the place holder is {8/10, 8/12, 8/14,
8/16}.
Also the data structures are generalized in order to support more
flexible factors instead of sticking with a fixed numerator(8).

(Cherry-picked from research-in-loop-filters branch on googlesource)