Add quarternary command line flag for optical flow refinement
A command line symbol called --enable-opfl-refine is added and has
the following values:

0: off, 1: switchable per block (default), 2: always applied in
compound average blocks, 3: auto (encoder can decide at frame level
which of the three option is used).

The frame wise decision criteria for the encoder is not included now.
It is left for future optimization.

STATS_CHANGED