Bugfix for global motion with 16bit internal.
Earlier, when 16bit internal was being used, the value of
'cpi->source->buf_8bit_valid' was not being reset to 0 on some cases.

Because of this, av1_downconvert_frame() was incorrectly assuming that
'cpi->source->y_buffer_8bit' has already been computed previously (from
16 bit 'cpi->source->y_buffer'), thereby using values 8-bit pixel values
computed from an earlier source frame.

We reset this flag appropriately to fix this bug.