Advanced MVD coding
In this commit, two advanced MVD coding methods are proposed. First, for NEW_NEARMV and NEAR_NEWMV mode, the precision of the MVD is proposed to depend on the associated class of motion vector difference and the magnitude of MVD. Second, a new inter coded mode, named as JOINT_NEWMV, is added to indicate whether the MVDs for two reference lists are jointly signaled. If the inter prediction mode is equal to JOINT_NEWMV mode, only one MVD, named as joint_delta_mv, is signaled and transmitted to the decoder, and the MVDs for reference list 0 and reference list 1 are derived from joint_delta_mv.