Correct temp buffersizes when superres is used
Prior to this fix, temporary buffers last_frame_uf and cnn_buffer
were at the coded size. However when superres is used they need to
be at the upscaled size to compare between LR and CNN. This patch
corrects the issue by adding a separate buffer for pre-CNN/LR and
correcting (and renaming) the size of the post CNN buffer.