Fix edge case in SSE4/AVX2 code for Wiener filter selection
The various av1_compute_stats[_highbd]_[sse4_1|avx2] functions
process two horizontal pixels at a time, but did not have any
code to handle the case where the restoration unit width is odd.
So, when this width was odd, the code was reading an extra pixel
beyond the restoration unit + 3-pixel border, and this was
leading to bad results.

This manifested as a mismatch between encoder results between the
SIMD-on and SIMD-off case:
https://gitlab.com/AOMediaCodec/avm/-/issues/42

This patch resolves this issue by adding explicit handling for
the last pixel column of an odd-width restoration unit. It also
updates the test cases for this code, to a form which would have
caught the original issue.

STATS_CHANGED