Update super-res models
New models are a little simpler that previous models in that they
use more 3x3 convolutions than 5x5.
Thse models use the gdvr-f architecture with 6 residue layers,
and an overall #parameters of about 3.5k and FLOPS per pixel
of 7K.
Validation error stats are about the same as the older models.