ERP: Fix prune_ref_frame_for_rect_partitions speed feature
This speed feature was broken by the ERP experiment because the sf only
applies when MB_MODE_INFO::partition is not PARTITION_NONE. But in the
ERP experiment, the partition is always PARTITION_NONE when
rd_pick_sb_mode is called, which disables the speed features.

In this fix, we check for the condition by checking whether any of the
HORZ/VERT subblock has been successfully searched.

Performance over 16 frames:
| SPD_SET | TESTSET | AVG_PSNR | OVR_PSNR |  SSIM   | DELTA TIME | DELTA MAX TIME |
| :-----: | :-----: | :------: | :------: | :-----: | :--------: | :------------: |
|    0    | av2_a3  | +0.111%  | +0.116%  | +0.134% |   -14.0%   |     -22.0%     |
|    0    | av2_a4  | +0.124%  | +0.133%  | +0.148% |   -16.0%   |     -19.9%     |
|    0    | av2_a5  | +0.309%  | +0.344%  | +0.379% |   -18.0%   |     -19.7%     |

STATS_CHANGED

Change-Id: I85ab5ec532c032586563f54271d1017ae92c5300