Deduplicate num_inliers information
Before this patch, the number of inliers found by the RANSAC algorithm
was stored in both num_inliers_per_motion[] and
params_by_motion[], and then these values were used
interchangeably.

As well as using more memory than necessary, this was also a source
of inconsistency between the RANSAC function which took integer
inputs and the one which took double inputs - the latter only
set num_inliers_per_motion.

This patch stores one copy of the inlier count in params_by_motion,
so that this count is kept alongside the inliers array, and resolves
the mentioned inconsistency.