FixedQp Offets setting: arf to not depend on stats
In fixed Qp offsets setting, the strength of the temporal filter for
arf generation should not depend on the stats. This patch makes the
strength of the filter depend only on #frames on either side suppored
by the arf frame.

This patch also fixes the issue of two parallel encoding for two GOPs
producing different outputs from a single encode, since the difference
was caused by the difference in stats in the last subGOP of a GOP.

STATS_CHANGED