SubGopSwitchingTestLarge: set kf_min_dist.
Earlier, there was no lower limit for minimum keyframe distance. So, it
was possible that encoder may choose to insert multiple scene-cut
keyframes. And consequently, it was possible that subgop config supplied
by the user never gets used: as GF interval in subgop config (e.g. 16)
may not match the GF interval selected by the encoder in any of the subgops.

Now, we limit minium keyframe distance to 65 to avoid this issue.