Fix issue 83: MV overwriting rule
When NRS is on and CONFIG_TMVP_IMPROVEMENT is off, the projected
motion vectors from some reference frames overwrite the existing
projected motion vectors, but some do not. It includes partial of
code changes from CONFIG_TMVP_IMPROVEMENT is on.

When NRS is on and CONFIG_TMVP_IMPROVEMENT is off, the later projected
motion vectors should always overwrite the existing projected motion
vectors to keep it be consistent with the code without NRS.