RST_MERGECOEFFS: support any ref in bank w/ delta
Prior to this patch RST_MERGECOEFFS only supported exact
matches from a dynamic bank. Now it supports any element
from the bank to be used as reference to code a delta.

The first element in the dynamic list still codes the exact
match case with a short 1-bit code, so that the current encoder
implementation of RST_MERGECOEFFS is consistent.

If there is no exact match, then the best reference is signaled
first followed by a delta from it.