Use above and left cctx types as contexts
Introduce spatial contexts for cctx types using the above and
left cctx types. In order to do so, the whole block area in
cctx_type_map associated with the current block is updated
whenever cctx type is updated. For skip blocks cctx type map
is always set to CCTX_NONE, such that cctx_type_map reflects
the actual cctx type everywhere.

Context is 0 if the above and left cctx types are of opposite
signs or both unavailable. Otherwise, context is 1 if there is
a positive angle cctx in the above or left block, 2 if there is
a negative angle cctx in these neighbors, and 0 otherwise.

Change-Id: I6446ed36147fae17af29f566db891b99f5862431