EXTENDED_WARP_PREDICTION: Fix bug with FLEX_MVRES
The EXTENDED_WARP_PREDICTION patches moved a line
  mbmi->motion_mode = SIMPLE_TRANSLATION;
within handle_inter_mode(), to make sure it happened before some
other code which it affected.

However, when rebasing, this accidentally moved the line
outside of a "#if !CONFIG_FLEX_MVRES" block. This caused a lot
of unintended consequences when FLEX_MVRES was enabled, resulting
in a small BDRATE impact but a huge encode slowdown (~25%).

This patch fixes that issue by moving the line in question
inside the #if !CONFIG_FLEX_MVRES block.