Fix frame buffer overflow in mismatch debug (#96)
The overflow is due to the fact that MAX_FRAME_BUF_NUM must 
be at least 1 plus maximum pyramid level, which is 6. It is
because that the frame buffers in the mismatch debug tool 
are suppose to store a key frame and all the frames from 
higher pyramid level in order to handle the decode latency.