Reduce allowed cctx type set to size 3
With CCTX_ADAPT_REDUCED_SET flag, only 3 cctx types will be
searched and signaled based on the above and left cctx types.

However, whenever skip_txfm is overwritten as 1 at the encoder
side, the cctx must be reset to CCTX_NONE, which may not be
optimal and causes a performance drop. Thus, this flag is turned
off by default now.

Change-Id: Ib4e7aba7ee7ba748564994702c193dc4720aa84b