Remove high bithdepth pointer conversion logic
This was needed in order to do pointer arithmetic while maintaining pixel
strides for both low and high bitdepth codepaths for the purpose of reducing
code duplication.
Since the low bitdeph path is gone, this conversion can be removed.

This is a follow up to 857e93fa9519ec2478eacd44f32a40ac05535b23.

Signed-off-by: James Almer <jamrial@videolan.org>