Correct indexing for wienerns stats in RUs
Introduces a parameter rest_unit_idx_in_rutile in rest unit worker
indicating the unit index within an Ru-tile, in addition to the
rest_unit_idx parameter indicating the unit index within the frame.
This allows for accessing stats correctly and conveniently from the
wienerns_stats vector.
Previously the indexing mechanism was incorrect, causing multi-tile
encoding to break.

As a TODO, we need to rename things consistently so that there is less
confusion in the future.