Correct q assignment in superres modes
The qp assignment in super-res modes was incorrect before.
TO fix this properly, this patch changes the meaning of the
fixed_qp_offsets[] array to indicate the offset ratio instead
of the absolute offset. Also renames this array to fixed_q_offsets[]
for correctness since it is not a qp offset but rather a q-step
offset.
This change is not expected to affect the baseline results, since
the double precision math remains pretty much the same.