Disable the TIP frame level mode with superres
Currently the TIP frame level mode is not designed to handle
superres since there is no provision for upscaling the tip_ref
buffer.

Until this support is added, disable the TIP frame level mode,
normatively as a bitstream restriction. That is if superres is
used the TIP_FRAME_AS_OUTPUT mode cannot be used.

This patch also corrects for the case where cpi->scaled_source is
not always the same as the cpi->source even when superres is not
used, and fixes a prior regression.