Fix scaling drift in superres plus TIP
When TIP is on, on the encode side, the TIP frame is created
first and then individual blocks reference pixels from the
tip ref buffer. However on the decoder side it is generated
on-the-fly. Therefore the blocks used for generating the tip ref
buffer on the encoder side are not necessarily aligned to the
blocks used on the decoder side. This leads to an issue with
scaled prediction due to a rounding operation in getting the
initial high res subpel location of pixels at the left end of a
block. The high resolution subpel locations may slowly drift over
a block leading to a mismatch on the encoder and decoder sides
since they have different block boundaries.

The fix for this is not straightforward. For any value in q4
precision to be scaled we change the scaling mechanism as follows:
first, compute a high resolution scaled location of the lower
closest 128 aligned boundary; then, the remaining offset is
calculated, and then scaled by a step size per suboel q4 level without
any rounding. We need to make sure step size in integer pixel domain
is exactly 16 times the step per subpel (q4) level to avoid drift.
In this manner the high resolution subpel offsets for any pixel
become independent of the actual boundary from where the convolve
started, as long as they are contained within a 128x128 superblock.

To make this work, we also need to ensure that the tip ref generation
blocks do not cross 128x128 boundaries.