Simplify and generalize ru tile wise processing
Removes unnecessary structures and functions, simplifies the
encoder side processing flow, and eliminates duplications. Also
the processiong model is now generalized covering the case of
multiple rows of RUs mapping to a single SB (when RUs are very small),
which was not supported before.

RusPerTileHelper is no longer necessary and is removed.