Fix a bug in tile by tile RU processing
RUs are offset by 8 pixels upwards to allow striping as per
the AV1 spec. However the last RU can be extended by 8 pixels
vertically beyond 3/2 times ru_size.

The previous code was not handling the extension correctly
causing rare crashes (only one in the AV2 CTC) depending on
frame height. This patch fixes the issue. All other cases
are invariant - hence the previous tests did not catch this
before.

A new SB level foreach function is used instead of reusing the tile
level function to handle a subtle difference in the the way offsets
are addressed, in order to avoid future confusion.