Select best features for flow estimation
When the initial feature detection pass finds many points, we
need to trim this down to a manageable number, determined by the
value MAX_CORNERS.

Previously, we just took the first MAX_CORNERS many points in raster
order, but this has issues on large frames, as it ends up only selecting
points along the top edge of the frame.

This patch changes the logic so that we select the sharpest corners.
Sharper corners work better as feature points, and also don't have the
bias toward the top of the frame which we had before.

STATS_CHANGED

Change-Id: Iab375cec1c8663ab45318aa82033d3b258581d5f