Start new expt on global MV w/ projected MV field
This commit starts a new experiment to explore global motion localized
within a superblock based on temporal projected motion fields.

Single reference temporal motion fields are generated for each
8x8 sub-block of a given frame for each reference frame.
This motion field is then used to generate a global affine model for the
reference localized within each superblock. An integrized least-squares
solver on the 8x8 motion field is used to generate the affine model that
the decoder is expected to derive at the SB level.

This SB level affine model will subsequently be explored for developing
new warp modes or extending existing ones.