Fix homography estimation and turn on tests
The perspective parameters in a homography need to be positive and
very small. That is sufficient for the models we are interested in
estimating from a normal visual scene.
Besides, the robustness of the homography estimation process improves
substantially with normalization, which has been fixed and turned on
by default now.
With these changes the tests pass.