Move get_rand_indices to random.h
The function get_rand_indices used in RANSAC is a generic function for
selecting k random numbers out of a range [0, n). So it may be useful
in other contexts in future. Therefore move it to random.h and rename
it to lcg_pick(), to help future patch authors find it.

Change-Id: I184eca0d52b9ddfb8191c0bd2c0cc3e8bc251b67