Temporarily revert segmentation changes
Undo an earlier change to the way the global motion segmentation map
is constructed. The original change works well on AV1 (including on
the AOM CTC), but it causes significant losses on AV2. The reasons
for this discrepancy are unclear, and will need to be investigated
further later.

As this is an encoder-only change, we will leave this out of the
IMPROVED_GLOBAL_MOTION submission for now and submit an appropriate
fix further down the line when one is ready.