Add conv_unet_attention guided models
This commit adds the conv_unet_attention guided and unguided models.
This includes changes to the quant sets to include two scaling
factors, changes to qp buckets for model selection, and padding to
allow for deeper models. 

All of the models in av1/tflite_models which start with 'qp' have been
replaced with the conv_unet_attention model (~1680 FLOPs unguided
and ~1720 FLOPs guided). 

The full set of changes is as follows:
  - Add guided and unguided conv_unet_attention model in intra_frame_model
  - Add padding to allow for deeper models
  - Add a second scaling factor, expand the quant sets
  - Add guided inter-frame models

Danilo