Add hooks for temporal LR filter propagation
Adds an experiment where the restoration info used for a frame is exposed
to RefCntBufffer (reference buffer) so that it is available to subsequent
frames to use as a predictor.

The current version simply uses the LR information in the primary reference
frame (available as cm->prev_frame->rest_info) to initialize the banks
for the current frame.