Add pixel-based warped parameter estimation
Adds CONFIG_INTERINTRA_THICK_BORDER to process and make available
a thicker border around an inter predictor for use in various new
tools.
Based on that this patch adds CONFIG_INTERINTRA_WARP experiment with a
preliminary rotzoom estimator based on pixels in the 4-pixel wide intra
region and a 4-pixel border around an inter predictor.