[encoder] Rework global motion ref selection
The current logic for selecting which ref frames to use for
global motion estimation is inherited from AV1, and is a
little messy. However, the NEW_REF_SIGNALING experiment has
re-ordered the ref frames so that the most promising refs come
first, which means that all we really need is a simple value
which tells the encoder to consider the first N ref frames.

Recent experiments on AV1 indicate that, for most frames, the
optimal balance between the prediction improvement from global
motion and the cost of the extra signaling is at 2 ref frames.
However, for high-quality frames, going up to 4 ref frames
does provide a small extra benefit. The new speed feature
is tuned accordingly.

STATS_CHANGED