CWG-E052: Fix bit depth logic in Gaussian elimination
The max bit width for intermediate results in Gaussian
elimination is supposed to be 32 in the E052 proposal,
but it is observed that the actual bit width can be 33.

This CL changes the shift logic in the pre-processing stage,
which makes the bit width 32 as proposed. There is a minor
BD-rate difference around -0.01, tested on 4875ca009f.

STATS_CHANGED

Change-Id: Id99593ef09ba566dbf3277c7d6c34f24dd132dd5